refactor to satisfy dependency-cruiser
